+++ examples/flow-builder-typescript/src/bot-config.ts
@@ -1,12 +1,9 @@
-import { BotConfigJSON } from '@botonic/core'
+import type { BotConfigJSON } from '@botonic/core'
+import { getToolsForBotConfig } from '@botonic/plugin-ai-agents'
 
 import { customTools } from './server/tools'
 
-const aiAgentTools =
-  customTools?.map(customTool => ({
-    name: customTool.name,
-    description: customTool.description,
-  })) || []
+const aiAgentTools = getToolsForBotConfig(customTools ?? [])
 
 /**
  * This is the configuration is shared with flow-builder-frontend.